+++ src/main/java/com/disqueprogrammer/app/trackerfinance/persistence/entity/Budget.java
@@ -3,6 +3,7 @@ package com.disqueprogrammer.app.trackerfinance.persistence.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -20,9 +21,9 @@ public class Budget {
 
     private String detail;
 
-    private double limitAmount;
+    private BigDecimal limitAmount;
 
-    private double usedAmount;
+    private BigDecimal usedAmount;
 
     private LocalDateTime dateBegin;
